+++ cmd/api_key.go
@@ -3,7 +3,7 @@ package cmd
 import (
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
-	"groundcover.com/pkg/api"
+	"groundcover.com/pkg/auth"
 	"groundcover.com/pkg/ui"
 )
 
@@ -13,7 +13,7 @@ var apiKeyCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		var err error
 
-		var apiKey api.ApiKey
+		var apiKey auth.ApiKey
 		if err = apiKey.Load(); err != nil {
 			return errors.Wrap(err, "failed to load api key")
 		}